Endpoints de CmpClientes readys

+++ BussinessLayer/Interfaces/ModuloCampaña/ICmpClienteRepository.cs
@@ -4,10 +4,10 @@ namespace BussinessLayer.Interfaces.ModuloCampaña
 {
     public interface ICmpClienteRepository
     {
-        Task<IEnumerable<CmpCliente>> GetAllAsync(int empresaId);
-        Task<CmpCliente?> GetByIdAsync(int id,int empresaId);
+        Task<IEnumerable<CmpCliente>> GetAllAsync(long empresaId);
+        Task<CmpCliente?> GetByIdAsync(int id,long empresaId);
         Task AddAsync(CmpCliente cliente);
         Task UpdateAsync(CmpCliente cliente);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
